Adding the org users query to core
