--- src/Api/AdminConsole/Queries/IOrganizationUserUserDetailsQuery.cs
@@ -1,6 +1,0 @@
-﻿namespace Api.AdminConsole.Queries;
-
-public interface IOrganizationUserUserDetailsQuery
-{
-    Task<IEnumerable<OrganizationUserUserDetailsQueryResponse>> GetOrganizationUserUserDetails(OrganizationUserUserDetailsQueryRequest request);
-}
