+++ src/Core/AdminConsole/OrganizationFeatures/OrganizationUsers/Interfaces/IOrganizationUserUserDetailsQuery.cs
@@ -1,0 +1,9 @@
+﻿using Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Requests;
+using Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Responses;
+
+namespace Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Interfaces;
+
+public interface IOrganizationUserUserDetailsQuery
+{
+    Task<IEnumerable<OrganizationUserUserDetailsQueryResponse>> GetOrganizationUserUserDetails(OrganizationUserUserDetailsQueryRequest request);
+}
