+++ src/Core/AdminConsole/OrganizationFeatures/OrganizationUsers/Requests/OrganizationUserUserDetailsQueryRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Requests;
+
+public class OrganizationUserUserDetailsQueryRequest
+{
+    public Guid OrganizationId { get; set; }
+    public bool IncludeGroups { get; set; } = false;
+    public bool IncludeCollections { get; set; } = false;
+}
